src: shrink line_offset in struct location to 4 bytes
line_offset of 2^32 bytes should be enough.

This requires the removal of the last_line field (in a previous patch) to
shrink struct expr to 112 bytes.

+++ include/nftables.h
@@ -158,8 +158,7 @@ struct location {
 	const struct input_descriptor		*indesc;
 	union {
 		struct {
-			off_t			line_offset;
-
+			unsigned int 		line_offset;
 			unsigned int		first_line;
 			unsigned int		first_column;
 			unsigned int		last_column;
